requirements.txt
about-time==3.1.1
aiohttp==3.8.3
aiosignal==1.3.1
alive-progress==2.4.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
discord-sentry-reporting==2.0.4
discord.py==2.1.0
dpytest==0.6.0
exceptiongroup==1.0.4
Flask==2.2.2
frozenlist==1.3.3
grapheme==0.6.0
idna==3.4
importlib-metadata==5.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.3
packaging==22.0
peewee==3.15.4
pip==21.3.1
pluggy==1.0.0
pycparser==2.21
pygit2==1.11.1
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
pytest-asyncio==0.20.3
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
sentry-sdk==1.11.1
setuptools==60.2.0
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
yarl==1.8.2
zipp==3.11.0
